+++ playground/window/src/main.c
@@ -1,14 +1,6 @@
-#include "window.h"
-
 #include "raykit.h"
-
-#include <stdio.h>
-
-#include "player/player.h"
 
 int main(void)
 {
-    printf("nadd: %d\n", number());
-    printf("add: %d\n", sum_add(2, 3));
-    return run();
+    return raykit_run();
 }
