--- playground/window/src/player/player.c
@@ -1,6 +1,0 @@
-#include "player.h"
-
-int number()
-{
-    return 1;
-}
