--- playground/window/src/player/player.h
@@ -1,3 +1,0 @@
-#pragma once
-
-int number();
