--- playground/window/src/window.h
@@ -1,1 +1,0 @@
-int run(void);
