+++ raykit/src/raykit.c
@@ -1,8 +1,9 @@
-#include "config.h"
-#include "renderer.h"
 #include "raylib.h"
+#include "raykit.h"
 
-int run(void)
+#include "window/renderer.h"
+
+int raykit_run(void)
 {
     InitWindow(CONFIG_SCREEN_WIDTH, CONFIG_SCREEN_HEIGHT, CONFIG_TITLE);
     SetTargetFPS(CONFIG_SCREEN_FPS);
