--- raykit/src/sum.c
@@ -1,6 +1,0 @@
-#include "raykit.h"
-
-int sum_add(int a, int b)
-{
-    return a + b;
-}
